% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{new_variable}
\alias{new_variable}
\title{Variable Container}
\usage{
new_variable(name, x, transform)
}
\arguments{
\item{name}{Character name of the variable}

\item{x}{A vector containing the variable data}

\item{transform}{An object created using \link{new_transform} that contains
information necessary to bin a variable.}
}
\value{
A list of the above arguments with a hist element that is a list of
all changes to the transform.
}
